Facebook iOS SDK 3.13.1
Summary:
    see https://developers.facebook.com/ios/change-log-3.x/

Upgrading:
    https://developers.facebook.com/docs/ios/upgrading/

+++ src/Login/FBSessionAppEventsState.h
@@ -26,7 +26,7 @@ extern NSString *const kFBAppEventIsImplicit;
 
 @property (readonly, retain) NSMutableArray *accumulatedEvents;
 @property (readonly, retain) NSMutableArray *inFlightEvents;
-@property (readwrite) int numSkippedEventsDueToFullBuffer;
+@property (readwrite) NSUInteger numSkippedEventsDueToFullBuffer;
 @property (readwrite) BOOL requestInFlight;
 
 - (void)addEvent:(NSDictionary *)eventDictionary
